- fix resize test

+++ test/kernels/cuda/test_cuda_resize.cc
@@ -11,7 +11,7 @@ TEST(Resize, Cuda_downsample_sizes_nearest) {
     Graph gCpu = make_ref<GraphObj>(runtime);
 
     auto input = gCpu->addTensor({1, 1, 2, 4}, DataType::Float32);
-    auto sizes = gCpu->addTensor({4}, DataType::UInt32);
+    auto sizes = gCpu->addTensor({4}, DataType::UInt64);
     gCpu->dataMalloc();
     input->copyin(vector<float>{1, 2, 3, 4, 5, 6, 7, 8});
     sizes->copyin(vector<uint32_t>{1, 1, 1, 3});
@@ -40,7 +40,7 @@ TEST(Resize, Cuda_upsample_sizes_nearest_notlarger) {
     Graph gCpu = make_ref<GraphObj>(runtime);
 
     auto input = gCpu->addTensor({1, 1, 2, 2}, DataType::Float32);
-    auto sizes = gCpu->addTensor({2}, DataType::UInt32);
+    auto sizes = gCpu->addTensor({2}, DataType::UInt64);
     gCpu->dataMalloc();
     input->copyin(vector<float>{1, 2, 3, 4});
     sizes->copyin(vector<uint32_t>{7, 8});
@@ -73,7 +73,7 @@ TEST(Resize, Cuda_upsample_sizes_nearest_notsmaller) {
     Graph gCpu = make_ref<GraphObj>(runtime);
 
     auto input = gCpu->addTensor({1, 1, 2, 2}, DataType::Float32);
-    auto sizes = gCpu->addTensor({2}, DataType::UInt32);
+    auto sizes = gCpu->addTensor({2}, DataType::UInt64);
     gCpu->dataMalloc();
     input->copyin(vector<float>{1, 2, 3, 4});
     sizes->copyin(vector<uint32_t>{7, 8});
@@ -106,7 +106,7 @@ TEST(Resize, Cuda_upsample_sizes_nearest_ceil_half_pixel) {
     Graph gCpu = make_ref<GraphObj>(runtime);
 
     auto input = gCpu->addTensor({1, 1, 4, 4}, DataType::Float32);
-    auto sizes = gCpu->addTensor({4}, DataType::UInt32);
+    auto sizes = gCpu->addTensor({4}, DataType::UInt64);
     gCpu->dataMalloc();
     input->copyin(
         vector<float>{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16});
@@ -143,7 +143,7 @@ TEST(Resize, Cuda_upsample_sizes_nearest_floor_align_corners) {
     Graph gCpu = make_ref<GraphObj>(runtime);
 
     auto input = gCpu->addTensor({1, 1, 4, 4}, DataType::Float32);
-    auto sizes = gCpu->addTensor({2}, DataType::UInt32);
+    auto sizes = gCpu->addTensor({2}, DataType::UInt64);
     gCpu->dataMalloc();
     input->copyin(
         vector<float>{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16});
@@ -180,7 +180,7 @@ TEST(Resize, Cuda_upsample_sizes_nearest_round_prefer_ceil_asymmetri) {
     Graph gCpu = make_ref<GraphObj>(runtime);
 
     auto input = gCpu->addTensor({1, 1, 4, 4}, DataType::Float32);
-    auto sizes = gCpu->addTensor({4}, DataType::UInt32);
+    auto sizes = gCpu->addTensor({4}, DataType::UInt64);
     gCpu->dataMalloc();
     input->copyin(
         vector<float>{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16});
@@ -421,7 +421,7 @@ TEST(Resize, Cuda_downsample_sizes_linear_pytorchhalfpixel) {
     Graph gCpu = make_ref<GraphObj>(runtime);
 
     auto input = gCpu->addTensor({1, 1, 4, 4}, DataType::Float32);
-    auto sizes = gCpu->addTensor({4}, DataType::UInt32);
+    auto sizes = gCpu->addTensor({4}, DataType::UInt64);
     gCpu->dataMalloc();
     input->copyin(
         vector<float>{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16});
@@ -453,7 +453,7 @@ TEST(Resize, Cuda_tf_crop_and_resize) {
     Graph gCpu = make_ref<GraphObj>(runtime);
 
     auto input = gCpu->addTensor({1, 1, 4, 4}, DataType::Float32);
-    auto sizes = gCpu->addTensor({4}, DataType::UInt32);
+    auto sizes = gCpu->addTensor({4}, DataType::UInt64);
     auto roi = gCpu->addTensor({8}, DataType::Float32);
     gCpu->dataMalloc();
     input->copyin(
@@ -490,7 +490,7 @@ TEST(Resize, Cuda_tf_crop_and_resize_axes_3_2) {
     Graph gCpu = make_ref<GraphObj>(runtime);
 
     auto input = gCpu->addTensor({1, 1, 4, 4}, DataType::Float32);
-    auto sizes = gCpu->addTensor({2}, DataType::UInt32);
+    auto sizes = gCpu->addTensor({2}, DataType::UInt64);
     auto roi = gCpu->addTensor({4}, DataType::Float32);
     gCpu->dataMalloc();
     input->copyin(
@@ -715,7 +715,7 @@ TEST(Resize, Cuda_downsample_sizes_cubic) {
     Graph gCpu = make_ref<GraphObj>(runtime);
 
     auto input = gCpu->addTensor({1, 1, 4, 4}, DataType::Float32);
-    auto sizes = gCpu->addTensor({4}, DataType::UInt32);
+    auto sizes = gCpu->addTensor({4}, DataType::UInt64);
     gCpu->dataMalloc();
     input->copyin(
         vector<float>{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16});
@@ -753,7 +753,7 @@ TEST(Resize, Cuda_upsample_sizes_cubic) {
     Graph gCpu = make_ref<GraphObj>(runtime);
 
     auto input = gCpu->addTensor({1, 1, 4, 4}, DataType::Float32);
-    auto sizes = gCpu->addTensor({4}, DataType::UInt32);
+    auto sizes = gCpu->addTensor({4}, DataType::UInt64);
     gCpu->dataMalloc();
     input->copyin(
         vector<float>{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16});
